+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Time and oracle call statistics
 - ANTLR4 export more stable
 - Dune installs `quotient_poirot`
+- Heuristic save can be disabled
 
 # v0.3
 - Dune installs `poirot` and `bnf2antlr4`